requirements.txt
fastapi
uvicorn
openai==0.28
jinja2
python-dotenv
gunicorn
python-multipart
